Introduce PluginIdentifier type to ensure we don't lookup plugins by object identities
